backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
psycopg2-binary
pydantic
pydantic-settings
email-validator
passlib[bcrypt]
python-multipart
pyjwt
pymysql
requests
